Bugfix: use `double` division and not integer division in `get_time`.
`SDL_GetTicks` returns ticks as `Uint32` so to convert them from
milliseconds to seconds as `double` they must be divided by `double`.

+++ code/ylikuutio/time/time.cpp
@@ -25,6 +25,6 @@ namespace yli::time
 {
     double get_time()
     {
-        return SDL_GetTicks() / 1000; // `SDL_GetTicks` return milliseconds.
+        return static_cast<double>(SDL_GetTicks()) / 1000.0; // `SDL_GetTicks` return milliseconds.
     }
 }
